+++ sel/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## sel-0.1.0.0
 
 * Consistently rename ciphertext-related types and functions [#182](https://github.com/haskell-cryptography/libsodium-bindings/pull/182)
+* Add support for additional data (AD) in `Sel.SecretKey.Stream` [#183](https://github.com/haskell-cryptography/libsodium-bindings/pull/183)
 
 ## sel-0.0.3.0